Remove helper package by extracting its content to separate packages (#800)
* Extract bridge package from helper.

Helper is a general name that may represent anything.
Extracting bridge module from it improves the code stucture a little by
having a smaller, focused package, that does what its name suggest.


* Move EventIsForThisNode function to node package


* Rename helper package to client

The helper package now contains only client.go.
Rename it to represent it's actual content.
